ACTIN-1320: move write to ExtractionResultWriter
So picked up by various test apps
